Add files via upload
This project demonstrates serial communication between an Arduino and a Raspberry Pi. The Arduino is programmed to control an LED while sending status messages like "LED ON" and "LED OFF" over a serial connection. A Python script on the Raspberry Pi reads these incoming messages and prints them to the terminal, showcasing how to interface the two different platforms